+++ dynamite/weight_solvers.py
@@ -165,6 +165,10 @@ class LegacyWeightSolver(WeightSolver):
         # prepare fortran input file for nnls
         self.copy_kinematic_data()
         self.create_fortran_input_nnls()
+        self.logger.info(f'{__class__.__name__} is DEPRECATED and '
+                         'will be removed in a future version of '
+                         f'DYNAMITE. Use weight solver type NNLS instead '
+                         f'of {__class__.__name__} if you can.')
 
     def copy_kinematic_data(self):
         """Copy kin data to infil/ direc
@@ -244,10 +248,7 @@ class LegacyWeightSolver(WeightSolver):
         nn_file.write(text)
         nn_file.close()
 
-    def solve(self,
-              orblib=None,
-              ignore_existing_weights=False,
-              delete_legacy_files=True):
+    def solve(self, orblib=None, ignore_existing_weights=False):
         """Main method to solve NNLS problem.
 
         Parameters
@@ -259,10 +260,6 @@ class LegacyWeightSolver(WeightSolver):
         ignore_existing_weights : bool
             If True, do not check for already existing weights and solve again.
             Default is False.
-        delete_legacy_files : bool
-            If True, legacy output files will be deleted after a successful
-            solve, otherwise they will remain in the model directory.
-            Default is True.
 
         Returns
         -------
@@ -336,26 +333,16 @@ class LegacyWeightSolver(WeightSolver):
             weights, chi2_tot, chi2_kin = \
                 self.get_weights_and_chi2_from_orbmat_file()
             chi2_kinmap = self.chi2_kinmap(weights)
-            # store legacy chi2 definitions (unused by DYNAMITE)
-            chi2_legacy, chi2_kin_legacy = self.read_chi2()
             results = table.Table()
             results['weights'] = weights
             results.meta = {'chi2_tot': chi2_tot,
                             'chi2_kin': chi2_kin,
-                            'chi2_kinmap': chi2_kinmap,
-                            'chi2_legacy': chi2_legacy,
-                            'chi2_kin_legacy': chi2_kin_legacy}
+                            'chi2_kinmap': chi2_kinmap}
             results.write(self.weight_file,
                           format='ascii.ecsv',
                           overwrite=True)
             # clean up
-            if delete_legacy_files:
-                for f_name in ['nn_aphist.out', 'nn_con.out',
-                               'nn_intrinsic_moments.out', 'nn_orb.out',
-                               'nn_orbmat.out', 'nn.in']:
-                    os.remove(self.direc_with_ml + f_name)
-                os.remove(self.fname_nn_kinem)
-                os.remove(self.fname_nn_nnls)
+            os.remove(self.direc_with_ml + 'nn_orbmat.out')
         return weights, chi2_tot, chi2_kin, chi2_kinmap
 
     def write_executable_for_weight_solver(self):